Switch from InternalsVisibleTo to reflection
